feat: pipeline false alarm checks, 64-bit chain counts, Markov NTLM8/9 tests, improved coverage model
- Pipeline false alarm checking in crackalack_lookup (launch/harvest pattern for GPU overlap)
- Widen chain count fields to uint64_t to support large tables (>4B chains)
- Add per-kernel GWS tuning (get_optimal_gws now takes kernel name)
- Improve coverage model in crackalack_plan with iterative merge accounting
- Add Markov NTLM8/NTLM9 chain test files and unit test registrations
- Update test_markov_accuracy.sh with train/test split and improved params
- Fix Metal kernel address space qualifiers and buffer argument handling
- Various kernel fixes for Markov fast-path (pspace_table, sorted_bigram)

+++ crackalack_rtc2rt.c
@@ -18,6 +18,7 @@
 #ifdef _WIN32
 #include <windows.h>
 #endif
+#include <inttypes.h>
 #include <stdio.h>
 
 #include "rtc_decompress.h"
@@ -26,7 +27,7 @@
 
 int main(int ac, char **av) {
   uint64_t *uncompressed_table = NULL;
-  unsigned int i = 0, num_chains = 0;
+  uint64_t i = 0, num_chains = 0;
   char *rtc_filename_input = NULL, *rt_filename_output = NULL;
   int ret = 0;
   FILE *f = NULL;
@@ -61,6 +62,6 @@ int main(int ac, char **av) {
 
   fclose(f);
 
-  printf("Successfully uncompressed %u chains in RTC file \"%s\" to RT file \"%s\".\n", num_chains, rtc_filename_input, rt_filename_output);
+  printf("Successfully uncompressed %"PRIu64" chains in RTC file \"%s\" to RT file \"%s\".\n", num_chains, rtc_filename_input, rt_filename_output);
   return 0;
 }
